backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pillow==11.0.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
google-genai==0.3.0
aiofiles==24.1.0
sqlalchemy==2.0.36
